+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- settings:
+  - Add `dump()` method on to print se
+  - Add `name` attribute on `SettingsDefinitionLoaderYaml` and
+    `RuntimeSettingsSiteLoaderIni` classes.
+  - Add `_origin` dict attribute on `RuntimeSettingsSection` to keep tracks of
+    origin of parameters values.
+
 ## [1.0.3] - 2024-08-30
 
 ### Added